Removes the HTML enum/namespace. Fixes a compilation error.
