+++ Classes/Private/Libxml2/Libxml2.swift
@@ -2,11 +2,11 @@ import Foundation
 
 enum Libxml2 {
     enum In {
-        typealias HTML = Libxml2.HTML
+        typealias HTML = Libxml2
     }
 
     enum Out {
-        typealias HTML = Libxml2.HTML
+        typealias HTML = Libxml2
     }
 
     enum HTML {
