+++ Classes/Private/NSAttributedString/Converters/HTMLNodeToNSAttributedString.swift
@@ -2,10 +2,10 @@ import Foundation
 
 class HMTLNodeToNSAttributedString: SafeConverter {
 
-    typealias ElementNode = Libxml2.HTML.ElementNode
-    typealias Node = Libxml2.HTML.Node
-    typealias RootNode = Libxml2.HTML.RootNode
-    typealias TextNode = Libxml2.HTML.TextNode
+    typealias ElementNode = Libxml2.ElementNode
+    typealias Node = Libxml2.Node
+    typealias RootNode = Libxml2.RootNode
+    typealias TextNode = Libxml2.TextNode
 
     /// The default font descriptor that will be used as a base for conversions.
     ///
@@ -184,7 +184,7 @@ class HMTLNodeToNSAttributedString: SafeConverter {
             let linkURL: String
 
             if let attributeIndex = node.attributes.indexOf({ $0.name == HTMLLinkAttribute.Href.rawValue }),
-                let attribute = node.attributes[attributeIndex] as? Libxml2.HTML.StringAttribute {
+                let attribute = node.attributes[attributeIndex] as? Libxml2.StringAttribute {
 
                 linkURL = attribute.value
             } else {
