+++ Example/Tests/Exporter/OutNodeConverterTests.swift
@@ -4,9 +4,9 @@ import libxml2
 
 class OutNodeConverterTests: XCTestCase {
 
-    typealias ElementNode = Libxml2.HTML.ElementNode
-    typealias Node = Libxml2.HTML.Node
-    typealias TextNode = Libxml2.HTML.TextNode
+    typealias ElementNode = Libxml2.ElementNode
+    typealias Node = Libxml2.Node
+    typealias TextNode = Libxml2.TextNode
 
     override func setUp() {
         super.setUp()
